Make participatesInLevel take a Long
The level lists are already Longs, so this keeps the generic extenders
conversion-free; the remaining Int/Long conversions move to the join
loops and the leapfrog implementers that store levels as Int.

+++ src/main/kotlin/org/hooray/algo/CombiJoin.kt
@@ -67,7 +67,7 @@ class CombiJoin (val extenders: List<CombiJoinExtender>, levels: Int): Join<Resu
     private val extenderSets : List<List<CombiJoinExtender>> = List(levels) { level ->
         val participants = mutableListOf<CombiJoinExtender>()
         for (extender in extenders) {
-            if (extender.participatesInLevel(level)) {
+            if (extender.participatesInLevel(level.toLong())) {
                 participants.add(extender)
             }
         }
