+++ src/main/kotlin/org/hooray/iterator/GenericNotPrefixExtender.kt
@@ -41,5 +41,5 @@ class GenericNotPrefixExtender(val children: List<PrefixExtender>, val levels: L
 
     override fun variableLevels(): List<Long> = levels
 
-    override fun participatesInLevel(level: Int) = level.toLong() == levels.last()
+    override fun participatesInLevel(level: Long) = level == levels.last()
 }
