+++ src/test/kotlin/org/hooray/iterator/AVLAndLeapfrogIndexTest.kt
@@ -190,7 +190,7 @@ class AVLAndLeapfrogIndexTest {
                 return levelData.size
             }
 
-            override fun participatesInLevel(level: Int): Boolean {
+            override fun participatesInLevel(level: Long): Boolean {
                 return level < levelData.size
             }
         }
@@ -283,7 +283,7 @@ class AVLAndLeapfrogIndexTest {
                 return levelData.size
             }
 
-            override fun participatesInLevel(level: Int): Boolean {
+            override fun participatesInLevel(level: Long): Boolean {
                 return level < levelData.size
             }
         }
