+++ src/test/kotlin/org/hooray/iterator/AVLOrLeapfrogIndexTest.kt
@@ -180,7 +180,7 @@ class AVLOrLeapfrogIndexTest {
                 return levelData.size
             }
 
-            override fun participatesInLevel(level: Int): Boolean {
+            override fun participatesInLevel(level: Long): Boolean {
                 return level < levelData.size
             }
         }
@@ -273,7 +273,7 @@ class AVLOrLeapfrogIndexTest {
                 return levelData.size
             }
 
-            override fun participatesInLevel(level: Int): Boolean {
+            override fun participatesInLevel(level: Long): Boolean {
                 return level < levelData.size
             }
         }
